backend/requirements.txt
fastapi
uvicorn[standard]
pandas
numpy
statsmodels
pydantic-settings
python-multipart
